% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dca.R
\name{compare_dca}
\alias{compare_dca}
\title{Compare DCA}
\usage{
compare_dca(..., as_plot_list = FALSE, .names = NULL)
}
\arguments{
\item{...}{Pass two bayesDCA fit objects with names as they should appear in plots'
legends (e.g. "Test one" = fit1, test2 = fit2).}

\item{as_plot_list}{If TRUE, returns list of ggplot2 plots (default is FALSE).}

\item{.names}{Optional character vector with names for objects (mostly for internal use).}
}
\value{
A figure of multiple plots
(\href{https://patchwork.data-imaginist.com/}{\code{patchwork}} object).
}
\description{
Compare DCA
}
\examples{

# Test 1: more specificity
data1 <- simulate_diagnostic_test_data(B = 1, N = 5000,
                                      true_p = 0.1, true_se = 0.9, true_sp = 0.8)
# Test 2: more sensitivity
data2 <- simulate_diagnostic_test_data(B = 1, N = 5000,
                                      true_p = 0.1, true_se = 0.95, true_sp = 0.7)

fit1 <- dca_binary_test(N = data1$N, d = data1$d, tp = data1$tp, tn = data1$tn)
fit2 <- dca_binary_test(N = data2$N, d = data2$d, tp = data2$tp, tn = data2$tn)
compare_dca("More Spec" = fit1, "More Sens" = fit2)

}
