% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dca_list.R
\name{compare_dca2}
\alias{compare_dca2}
\title{Plot BayesDCAList comparison}
\usage{
compare_dca2(
  obj,
  models_or_tests,
  colors = NULL,
  labels = NULL,
  plot_list = FALSE,
  ...
)
}
\arguments{
\item{obj}{BayesDCAList object}

\item{colors}{Named vector with color for each model or test. If provided
for a subset of models or tests, only that subset will be plotted.}

\item{labels}{Named vector with label for each model or test.}

\item{plot_list}{If TRUE, returns a list of separate ggplot objects.}
}
\value{
A patchwork/ggplot object or a list of ggplot objects.
}
\description{
Plot BayesDCAList comparison
}
\examples{
fit <- dca_list(PredModelData, cores = 4)
compare_dca2(fit)
compare_dca2(fit, models_or_tests = "binary_test")
}
