% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dca.R
\name{compare_threshold_posterior}
\alias{compare_threshold_posterior}
\title{Compare posterior distributions at specific thresholds}
\usage{
compare_threshold_posterior(..., thr, .names = NULL, data_only = FALSE)
}
\arguments{
\item{...}{Pass up to two bayesDCA fit objects with names as they should appear in plots'
legends (e.g. "Test one" = fit1, test2 = fit2). First object appears in y axis.}

\item{thr}{Threshold to use - should have been computed in the first object.}

\item{.names}{Optional character vector with names for objects (mostly for internal use).}

\item{data_only}{If TRUE, returns data for ggplot objects (default is FALSE).}
}
\value{
A ggplot object.
}
\description{
Compare posterior distributions at specific thresholds
}
