% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dca.R
\name{dca_binary_test}
\alias{dca_binary_test}
\title{Bayesian Decision Curve Analysis for Binary Test}
\usage{
dca_binary_test(
  N,
  d,
  tp,
  tn,
  thresholds = seq(0, 0.5, 0.01),
  keep_fit = FALSE,
  keep_draws = TRUE,
  prior_p = c(1, 1),
  prior_se = c(1, 1),
  prior_sp = c(1, 1),
  summary_probs = c(0.025, 0.975),
  refresh = 0,
  ...
)
}
\arguments{
\item{N}{Integer specifying total number of samples (i.e., number of participants).}

\item{d}{Diseased: integer specifying total number of diseased persons or outcome cases.}

\item{tp}{True Positives: integer specifying total number of diseased persons or cases correctly
identified as such by the binary test (positive tests among cases). \code{tp} must be at most \code{d}.}

\item{tn}{True Negatives: integer specifying total number of non-diseased persons or non-cases correctly
identified as such by the diagnostic test or prediction model (negative tests among non-cases). \code{tn} must be at most \code{N - d}.}

\item{thresholds}{Numeric vector with probability thresholds with which
the decision curve should be computed (default is \code{seq(0, 0.5, 0.01)}).}

\item{keep_fit}{Logical indicating whether to keep \href{https://mc-stan.org/rstan/reference/stanfit-class.html}{\code{stanfit}} object in
the output (default FALSE).}

\item{prior_p, prior_se, prior_sp}{Numeric vector with shapes for
Beta(alpha, beta) prior used for p, Se, and Sp. Default is c(1, 1) for all.}

\item{refresh}{Control verbosity of \href{https://mc-stan.org/rstan/reference/stanmodel-method-sampling.html}{\code{rstan::sampling}} (check its help
page for details).}

\item{...}{Arguments passed to \href{https://mc-stan.org/rstan/reference/stanmodel-method-sampling.html}{\code{rstan::sampling}} (e.g. iter, chains).}
}
\value{
An object of class \code{DiagTestDCA}
}
\description{
Bayesian Decision Curve Analysis for Binary Test
}
\examples{

fit <- dca_binary_test(N = 500, d = 83, tp = 77, tn = 378)
plot(fit)
}
