% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dca_list.R
\name{dca_list}
\alias{dca_list}
\title{Bayesian Decision Curve Analysis for a list of Predictive Models and Binary tests}
\usage{
dca_list(
  .data,
  thresholds = seq(0.01, 0.5, 0.01),
  keep_draws = TRUE,
  keep_fit = FALSE,
  prior_p = NULL,
  prior_se = NULL,
  prior_sp = NULL,
  summary_probs = c(0.025, 0.975),
  refresh = 0,
  ...
)
}
\arguments{
\item{.data}{A data.frame with an \code{outcomes} column (0 or 1 for each individual)
and one or more columns with predicted probabilities from each of desired list
of predictive models, or with 0 or 1 indicator from each of desired list of
binary tests.}

\item{thresholds}{Numeric vector with probability thresholds with which
the net benefit should be computed (default is \code{seq(0, 0.5, 0.01)}).}

\item{keep_draws}{Logical indicating whether to keep posterior
draws from \code{stanfit} object (default is TRUE).}

\item{keep_fit}{Logical indicating whether to keep \code{stanfit} in
the output (default is FALSE).}

\item{prior_p, prior_se, prior_sp}{Non-negative shape values for
Beta(alpha, beta) priors used for p, Se, and Sp, respectively.
Default is uniform prior for all parameters - Beta(1, 1).
A single vector of the form \code{c(a, b)} can be provided for each.}

\item{refresh}{Control verbosity of \code{rstan::sampling} (check its help
page for details).}

\item{...}{Arguments passed to \code{rstan::sampling} (e.g. iter, chains).}
}
\value{
An object of class \code{DCAList}
}
\description{
Estimate decision curves for a list of predictive models and/or
binary tests all at once. Necessary to make comparative inferences across
multiple models or tests
using their corresponding posterior draws.
}
\examples{
data(PredModelData)
head(PredModelData)
fit <- dca_list(.data = PredModelData)
plot(fit)
}
