% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dca_surv.R
\name{dca_surv_pem}
\alias{dca_surv_pem}
\title{Bayesian Decision Curve Analysis for
Survival outcomes (deprecated)}
\usage{
dca_surv_pem(
  .data,
  prediction_time,
  thresholds = seq(0, 0.5, length = 51),
  keep_draws = TRUE,
  keep_fit = FALSE,
  summary_probs = c(0.025, 0.975),
  cutpoints = NULL,
  prior_scaling_factor = 1/5,
  min_events = 10,
  prior_means = NULL,
  prior_only = FALSE,
  prior_anchor = c("prediction_time", "median"),
  keep_prior = FALSE,
  iter = 2000,
  refresh = 0,
  ...
)
}
\value{
An object of class \code{BayesDCASurv}
}
\description{
Bayesian Decision Curve Analysis for
Survival outcomes (deprecated)
}
\examples{
data(dca_survival_data)
fit <- dca_surv_pem(dca_survival_data, prediction_time = 1, cores = 4)
plot(fit)
}
\keyword{internal}
