% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dca_surv.R
\name{dca_surv_weibull2}
\alias{dca_surv_weibull2}
\title{Bayesian Decision Curve Analysis
for Survival outcomes (Weibull)}
\usage{
dca_surv_weibull2(
  .data,
  prediction_time,
  thresholds = seq(0, 0.5, length = 51),
  keep_draws = TRUE,
  keep_fit = FALSE,
  summary_probs = c(0.025, 0.975),
  positivity_prior = c(1, 1),
  prior_scale_alpha = 1,
  prior_scale_sigma = 100,
  prior_only = FALSE,
  prediction_time_scaling = TRUE,
  iter = 4000,
  refresh = 0,
  ...
)
}
\value{
An object of class \code{BayesDCASurv}
}
\description{
Bayesian Decision Curve Analysis
for Survival outcomes (Weibull)
}
\examples{
data(dca_survival_data)
fit <- dca_surv(dca_survival_data, prediction_time = 1, cores = 4)
plot(fit)
}
