% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dca_survival_data}
\alias{dca_survival_data}
\title{Survival Data for Bayesian DCA.}
\format{
A data frame with 500 rows and 3 variables:
\describe{
\item{outcomes}{survival outcomes (\code{survival::Surv} object)}
\item{model_predictions}{predicted probabilities of event (at time 1).}
\item{binary_test}{binary prognostic test results (0 or 1, int.)}
...
}
}
\source{
\url{https://giulianonetto.github.io/bayesdca/}
}
\usage{
dca_survival_data
}
\description{
A dataset containing the the time-to-event \code{outcomes} (a \code{survival::Surv} object),
the predictions from a prognostic model, and the results from a binary
prognostic test. Based on \code{dcurves::df_surv}. The intended time horizon for prediction
is exactly 1.
}
\keyword{datasets}
