% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Docking2021Data}
\alias{Docking2021Data}
\title{Survival data from Docking, 2021 - Figures 3B-3E}
\format{
A data frame with 635 rows and 5 variables:
\describe{
\item{fig_id}{Character indicating corresponding figure in the original paper.}
\item{aps_lasso_score}{APS value}
\item{aps_lasso_class}{APS tercile computed within study cohort.}
\item{survival_event}{Death is 1, censored/slive is 0.}
\item{survival_weeks}{Overall survival in weeks.}
...
}
}
\source{
\url{https://giulianonetto.github.io/bayesdca/}
}
\usage{
Docking2021Data
}
\description{
A dataset containing the survival data from \href{https://www.nature.com/articles/s41467-021-22625-y}{Docking, 2021}.
Dataset includes selected data used for Figures 3B to 3E.
Observations with missing survival time (in weeks)
or missing survival events were excluded.
AML prognostic score (APS) values were kept as an example of continuous biomarker.
}
\keyword{datasets}
