% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dca.R
\name{.dca_stan}
\alias{.dca_stan}
\title{Fit Bayesian Decision Curve Analysis using Stan}
\usage{
.dca_stan(
  N,
  d,
  tp,
  tn,
  B = NULL,
  thresholds = seq(0, 0.5, 0.01),
  model_type = "ic",
  prior_p = c(1, 1),
  prior_se = c(1, 1),
  prior_sp = c(1, 1),
  refresh = 0,
  ...
)
}
\arguments{
\item{N}{Integer specifying total number of samples (i.e., participants).}

\item{d}{Diseased: integer specifying total number of diseased persons.}

\item{tp}{True Positives: integer specifying total number of diseased persons correctly
identified as such by the diagnostic test of prediction model.}

\item{tn}{True Negatives: integer specifying total number of diseased persons correctly
identified as such by the diagnostic test of prediction model.}

\item{thresholds}{Numeric vector with probability thresholds with which
the net benefit should be computed (default is \code{seq(0, 0.5, 0.01)}).}

\item{model_type}{Character indicating model type. Either "correlated"
(default) or "independent".}

\item{prior_p, prior_se, prior_sp}{Numeric vector with shapes for
Beta(alpha, beta) prior used for p, Se, and Sp. Default is c(1, 1) for all.}

\item{refresh}{Control verbosity of \code{rstan::sampling} (check its help
page for details).}

\item{...}{Arguments passed to \code{rstan::sampling} (e.g. iter, chains).}
}
\value{
An object of class \code{stanfit} returned by \code{rstan::sampling}
}
\description{
Fit Bayesian Decision Curve Analysis using Stan
}
