% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dca_surv.R
\name{.dca_stan_surv2}
\alias{.dca_stan_surv2}
\title{Fit Bayesian Decision Curve Analysis using Stan for survival outcomes (hierarchical model)}
\usage{
.dca_stan_surv2(
  n_thr,
  n_models_or_tests,
  n_intervals,
  thresholds,
  time_exposed_prediction,
  deaths,
  time_exposed,
  posterior_alpha0,
  posterior_beta0,
  prior_only,
  pos_post1,
  pos_post2,
  other_models_indices,
  iter = 4000,
  refresh = 0,
  ...
)
}
\arguments{
\item{refresh}{Control verbosity of \href{https://mc-stan.org/rstan/reference/stanmodel-method-sampling.html}{\code{rstan::sampling}}.}

\item{...}{Arguments passed to \href{https://mc-stan.org/rstan/reference/stanmodel-method-sampling.html}{\code{rstan::sampling}} (e.g. iter, chains).}
}
\value{
An object of class \href{https://mc-stan.org/rstan/reference/stanfit-class.html}{\code{stanfit}} returned by \href{https://mc-stan.org/rstan/reference/stanmodel-method-sampling.html}{\code{rstan::sampling}} (e.g. iter, chains)
}
\description{
Fit Bayesian Decision Curve Analysis using Stan for survival outcomes (hierarchical model)
}
\keyword{internal}
