% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dca_surv.R
\name{.dca_stan_surv_pem}
\alias{.dca_stan_surv_pem}
\title{Fit Bayesian Decision Curve Analysis
using Stan for survival outcomes}
\usage{
.dca_stan_surv_pem(
  n_thr,
  n_models_or_tests,
  n_intervals,
  thresholds,
  time_exposed,
  posterior_alpha,
  posterior_beta,
  posterior_alpha0,
  posterior_beta0,
  pos_post1,
  pos_post2,
  other_models_indices,
  iter = 4000,
  refresh = 0,
  ...
)
}
\arguments{
\item{refresh}{Control verbosity of \href{https://mc-stan.org/rstan/reference/stanmodel-method-sampling.html}{\code{rstan::sampling}}.  # nolint}

\item{...}{Arguments passed to \href{https://mc-stan.org/rstan/reference/stanmodel-method-sampling.html}{\code{rstan::sampling}} (e.g. iter, chains). # nolint}
}
\value{
An object of class \href{https://mc-stan.org/rstan/reference/stanfit-class.html}{\code{stanfit}} returned by \href{https://mc-stan.org/rstan/reference/stanmodel-method-sampling.html}{\code{rstan::sampling}} (e.g. iter, chains)  # nolint
}
\description{
Fit Bayesian Decision Curve Analysis
using Stan for survival outcomes
}
\keyword{internal}
