% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dca_surv.R
\name{.dca_stan_surv_weibull2}
\alias{.dca_stan_surv_weibull2}
\title{Fit Bayesian Decision Curve Analysis
using Stan for survival outcomes (Weibull model)}
\usage{
.dca_stan_surv_weibull2(
  sample_size,
  n_thr,
  n_models,
  times,
  events,
  predictions,
  predictions_marginal,
  prediction_time,
  thresholds,
  other_models_indices,
  shape_prior,
  scale_prior,
  shape_prior_pars,
  scale_prior_pars,
  positivity_prior_pars,
  prior_only,
  iter,
  refresh,
  ...
)
}
\arguments{
\item{refresh}{Control verbosity of
\href{https://mc-stan.org/rstan/reference/stanmodel-method-sampling.html}{\code{rstan::sampling}}.  # nolint}

\item{...}{Arguments passed to
\href{https://mc-stan.org/rstan/reference/stanmodel-method-sampling.html}{\code{rstan::sampling}} (e.g. iter, chains).  # nolint}
}
\value{
An object of class
\href{https://mc-stan.org/rstan/reference/stanfit-class.html}{\code{stanfit}} returned by \href{https://mc-stan.org/rstan/reference/stanmodel-method-sampling.html}{\code{rstan::sampling}} (e.g. iter, chains)  # nolint
}
\description{
Fit Bayesian Decision Curve Analysis
using Stan for survival outcomes (Weibull model)
}
\keyword{internal}
