% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dca.R
\name{.extract_dca_draws}
\alias{.extract_dca_draws}
\title{Get posterior draws from DCA stanfit}
\usage{
.extract_dca_draws(fit, model_or_test_names)
}
\arguments{
\item{fit}{A stanfit object.}

\item{model_or_test_names}{Vector of names of models or binary tests under assessment.}
}
\description{
Get posterior draws from DCA stanfit
}
