% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dca_surv.R
\name{.extract_dca_surv_summary}
\alias{.extract_dca_surv_summary}
\title{Extract posterior summaries for \code{BayesDCASurv} object}
\usage{
.extract_dca_surv_summary(fit, summary_probs, thresholds, model_or_test_names)
}
\value{
An object of class \code{BayesDCASurv}
}
\description{
Extract posterior summaries for \code{BayesDCASurv} object
}
\keyword{internal}
