% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_prior_parameters}
\alias{.get_prior_parameters}
\title{Get priors for Bayesian DCA}
\usage{
.get_prior_parameters(
  thresholds,
  threshold_varying_prior = FALSE,
  n_strategies = NULL,
  prior_p = NULL,
  prior_se = NULL,
  prior_sp = NULL,
  ignorance_region_cutpoints = c(0.25, 0.75) * max(thresholds),
  min_sens_prior_mean = 0.01,
  max_sens_prior_mean = 0.99,
  max_sens_prior_sample_size = 5,
  ignorance_region_mean = 0.5,
  ignorance_region_sample_size = 2,
  prev_prior_mean = 0.5,
  prev_prior_sample_size = 2
)
}
\arguments{
\item{thresholds}{Vector of decision thresholds.}

\item{shift}{Scalar controlling height of prior
Specificity curve. Only used if \code{constant=FALSE}.}

\item{slope}{Scalar controlling shape of prior
Specificity curve. Only used if \code{constant=FALSE}.}

\item{min_mean_se, min_mean_sp, max_mean_se, max_mean_se}{Minimum}

\item{prior_sample_size}{Prior sample size of strength.}

\item{min_prior_sample_size}{Minimum prior sample size or strength.}

\item{max_prior_sample_size}{Maximum prior sample size or strength.}

\item{slope_prior_sample_size}{Rate of change in prior
sample size or strength.
and maximum prior mean for sensitivity (se) and specificity (sp).}
}
\description{
Get priors for Bayesian DCA
}
