% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dca_surv.R
\name{.get_relevant_pars}
\alias{.get_relevant_pars}
\title{Get relevant parameters to parse from Stan output (BayesDCASurv)}
\usage{
.get_relevant_pars()
}
\description{
Get relevant parameters to parse from Stan output (BayesDCASurv)
}
