% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dca.R
\name{.get_thr_data_list}
\alias{.get_thr_data_list}
\title{Get Threshold Performance Data for DCA list}
\usage{
.get_thr_data_list(
  .data,
  thresholds = seq(0.01, 0.5, 0.01),
  prior_only = FALSE
)
}
\arguments{
\item{.data}{A data.frame with an \code{outcomes} column (0 or 1 for each individual)
and one or more columns with predicted probabilities from each of desired list
of predictive models, or with 0 or 1 indicator from each of desired list of
binary tests.}

\item{thresholds}{Decision thresholds.}

\item{prior_only}{If TRUE, returns prior only (ignores any data).}
}
\description{
Get Threshold Performance Data for DCA list
}
\keyword{internal}
