% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_threshold_varying_prior_parameters}
\alias{.get_threshold_varying_prior_parameters}
\title{Get threshold-varying priors for Bayesian DCA}
\usage{
.get_threshold_varying_prior_parameters(
  thresholds,
  n_strategies,
  ignorance_region_cutpoints = c(0.25, 0.75) * max(thresholds),
  min_sens_prior_mean = 0.01,
  max_sens_prior_mean = 0.99,
  max_sens_prior_sample_size = 5,
  ignorance_region_mean = 0.5,
  ignorance_region_sample_size = 2,
  prev_prior_mean = 0.5,
  prev_prior_sample_size = 2
)
}
\description{
Get threshold-varying priors for Bayesian DCA
}
\keyword{internal}
