% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{evpi}
\alias{evpi}
\title{Compute Expected Value of Perfect Information (EVPI)}
\usage{
evpi(thresholds, ...)
}
\arguments{
\item{...}{matrices of dimension n_draws * n_thresholds
containing posterior net benefit}

\item{obj}{BayesDCAList or BayesDCASurv object}
}
\description{
Compute Expected Value of Perfect Information (EVPI)
}
\keyword{internal}
