% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{evpi_old}
\alias{evpi_old}
\title{Compute Expected Value of Perfect Information (EVPI)}
\usage{
evpi_old(obj, models_or_tests = NULL, type = c("best", "useful", "pairwise"))
}
\arguments{
\item{obj}{BayesDCAList object}

\item{models_or_tests}{Character vector with models or tests to compare. If null, compares either first two in \code{obj$model_or_test_names} or the first one against Treat all/none (if only one available).}
}
\description{
Compute Expected Value of Perfect Information (EVPI)
}
\keyword{internal}
