% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dca.R
\name{extract_dca_summary}
\alias{extract_dca_summary}
\title{Get summary from BayesDCA fit}
\usage{
extract_dca_summary(
  fit,
  model_or_test_names,
  summary_probs = c(0.025, 0.975),
  thresholds = seq(0.01, 0.5, 0.01)
)
}
\arguments{
\item{fit}{A stanfit object.}

\item{model_or_test_names}{Vector of names of models or binary tests under assessment.}

\item{summary_probs}{Numeric vector giving probabilities for credible interval.}

\item{thresholds}{Vector of thresholds for DCA.}
}
\description{
Get summary from BayesDCA fit
}
