% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dca_list.R
\name{get_colors_and_labels}
\alias{get_colors_and_labels}
\title{Get colors and labels for BayesDCA plots}
\usage{
get_colors_and_labels(
  obj,
  colors = NULL,
  labels = NULL,
  models_or_tests = NULL
)
}
\arguments{
\item{obj}{BayesDCAList object}

\item{colors}{Named vector with color for each model or test. If provided
for a subset of models or tests, only that subset will be plotted.}

\item{labels}{Named vector with label for each model or test.}
}
\description{
Get colors and labels for BayesDCA plots
}
