% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_cutpoints}
\alias{get_cutpoints}
\title{Get cutpoints for survival estimation}
\usage{
get_cutpoints(.prediction_time, .event_times)
}
\arguments{
\item{.prediction_time}{time point at which event is predicted to happen}

\item{.event_times}{times of observed events (non-censored)}
}
\description{
Get cutpoints for survival estimation
}
