% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_death_pseudo_counts}
\alias{get_death_pseudo_counts}
\title{Get death pseudo-counts}
\usage{
get_death_pseudo_counts(
  .prediction_data,
  .surv_data,
  .cutpoints,
  .models_or_tests,
  .thresholds
)
}
\arguments{
\item{.prediction_data}{Contains prognostic model predictions}

\item{.surv_data}{Contains observed time to event
(\code{.time} column) and event indicator (\code{.status} column)}

\item{.cutpoints}{Survival cutpoints.}

\item{.models_or_tests}{Names for models or tests.}

\item{.thresholds}{DCA thresholds.}

\item{.prior_scaling_factor}{Prior
alpha = (0.69/median_surv)*.prior_scaling_factor,
Prior beta = .prior_scaling_factor.}
}
\description{
Get death pseudo-counts
}
\keyword{internal}
