% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dca.R
\name{get_delta_plot_data_binary}
\alias{get_delta_plot_data_binary}
\title{Get delta plot data for binary bayesDCA}
\usage{
get_delta_plot_data_binary(
  obj,
  models_or_tests = NULL,
  type = c("best", "useful", "pairwise")
)
}
\arguments{
\item{obj}{BayesDCA object}

\item{models_or_tests}{Character vector with subset of decision strategies.}

\item{type}{One of "best", "useful", or "pairwise".}
}
\value{
A ggplot object.
}
\description{
Get delta plot data for binary bayesDCA
}
\keyword{internal}
