% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dca_surv.R
\name{get_delta_plot_data_surv}
\alias{get_delta_plot_data_surv}
\title{Get delta plot data for binary bayesDCA}
\usage{
get_delta_plot_data_surv(
  obj,
  strategies = NULL,
  type = c("best", "useful", "pairwise"),
  labels = NULL
)
}
\arguments{
\item{obj}{BayesDCA object}

\item{strategies}{Character vector with subset of decision strategies.}

\item{type}{One of "best", "useful", or "pairwise".}

\item{labels}{Named vector with label for each model or test.}
}
\value{
A ggplot object.
}
\description{
Get delta plot data for binary bayesDCA
}
\keyword{internal}
