% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_events_per_interval}
\alias{get_events_per_interval}
\title{Get events per intervals defined by set of cutpoints}
\usage{
get_events_per_interval(.cutpoints, .event_times)
}
\arguments{
\item{.cutpoints}{cutpoints defining intervals}

\item{.event_times}{times of observed events (non-censored)}
}
\description{
Get events per intervals defined by set of cutpoints
}
\keyword{internal}
