% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_positivity_posterior_parameters}
\alias{get_positivity_posterior_parameters}
\title{Get posterior parameters for positivity probability}
\usage{
get_positivity_posterior_parameters(
  .prediction_data,
  .thresholds,
  .prior_shape1 = 0.5,
  .prior_shape2 = 0.5
)
}
\arguments{
\item{.prediction_data}{Contains prognostic model predictions}

\item{.thresholds}{DCA thresholds.}

\item{.prior_shape1}{Shape 1 for beta prior.}

\item{.prior_shape2}{Shape 2 for beta prior.}
}
\description{
Get posterior parameters for positivity probability
}
