% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_prior_sample_size}
\alias{get_prior_sample_size}
\title{Get threshold-specific prior sample size or strength}
\usage{
get_prior_sample_size(
  thresholds,
  min_prior_sample_size = 20,
  max_prior_sample_size = 100,
  slope_prior_sample_size = 300
)
}
\arguments{
\item{thresholds}{Decision thresholds}

\item{min_prior_sample_size}{Minimum prior sample size or strength.}

\item{max_prior_sample_size}{Maximum prior sample size or strength.}

\item{slope_prior_sample_size}{Rate of change in
prior sample size or strength.}
}
\description{
Get threshold-specific prior sample size or strength
}
\keyword{internal}
