% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_survival_posterior_parameters}
\alias{get_survival_posterior_parameters}
\title{Get posterior parameters for survival model}
\usage{
get_survival_posterior_parameters(
  .prediction_data,
  .surv_data,
  .cutpoints,
  .models_or_tests,
  .thresholds,
  .prior_scaling_factor,
  .prediction_time,
  .prior_anchor = c("median", "prediction_time"),
  .prior_only = FALSE,
  .prior_means = NULL
)
}
\arguments{
\item{.prediction_data}{Contains prognostic model predictions}

\item{.surv_data}{Contains observed time to event
(\code{.time} column) and event indicator (\code{.status} column)}

\item{.cutpoints}{Survival cutpoints.}

\item{.models_or_tests}{Names for models or tests.}

\item{.thresholds}{DCA thresholds.}

\item{.prior_scaling_factor}{Prior}

\item{.prior_only}{If TRUE, will sample from prior only.
alpha = (0.69/median_surv)*.prior_scaling_factor,
Prior beta = .prior_scaling_factor.}
}
\description{
Get posterior parameters for survival model
}
\keyword{internal}
