% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_survival_time_exposed}
\alias{get_survival_time_exposed}
\title{Get time exposed within each interval for a given prediction time}
\usage{
get_survival_time_exposed(.prediction_time, .cutpoints)
}
\arguments{
\item{.prediction_time}{Time for event prediction}

\item{.cutpoints}{Cutpoints for constant hazard interval}
}
\description{
Get time exposed within each interval for a given prediction time
}
