% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dca.R
\name{get_thr_data}
\alias{get_thr_data}
\title{Get Threshold Performance Data}
\usage{
get_thr_data(
  df,
  outcome = "outcomes",
  prediction = "predictions",
  thresholds = seq(0, 0.5, 0.02)
)
}
\arguments{
\item{df}{Input data.frame}

\item{outcome}{Character indicating the column with binary outcome (as 0/1s).}

\item{prediction}{Character indicating the column with predicted probabilities.}
}
\description{
Get Threshold Performance Data
}
