% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dca.R
\name{get_thr_data_list}
\alias{get_thr_data_list}
\title{Get Threshold Performance Data for DCA list}
\usage{
get_thr_data_list(.data, thresholds = seq(0.01, 0.5, 0.01))
}
\arguments{
\item{outcomes}{Integer vector (0 or 1) with binary outcomes.}

\item{predictions}{Numeric vector with predicted probabilities.}
}
\description{
Get Threshold Performance Data for DCA list
}
