% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.BayesDCA}
\alias{plot.BayesDCA}
\title{Plot BayesDCA}
\usage{
\method{plot}{BayesDCA}(
  obj,
  models_or_tests = NULL,
  colors = NULL,
  labels = NULL,
  raw_values = NULL,
  raw_values_label = "Biomarker threshold",
  ...
)
}
\arguments{
\item{obj}{BayesDCA object}

\item{models_or_tests}{Character vector with
models or tests to compare. If null, compares
either first two in \code{obj$model_or_tests}
or the first one against Treat all/none
(if only one available).}

\item{colors}{Named vector with color for each model or test. If provided
for a subset of models or tests, only that subset will be plotted.}

\item{labels}{Named vector with label for each model or test.}
}
\description{
Plot BayesDCA
}
