% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dca_list.R
\name{plot.BayesDCAList}
\alias{plot.BayesDCAList}
\title{Plot BayesDCAList}
\usage{
\method{plot}{BayesDCAList}(obj, colors = NULL, labels = NULL, models_or_tests = NULL, ...)
}
\arguments{
\item{obj}{BayesDCAList object}

\item{colors}{Named vector with color for each model or test. If provided
for a subset of models or tests, only that subset will be plotted.}

\item{labels}{Named vector with label for each model or test.}

\item{models_or_tests}{Character vector specifying subset of models
or tests to be plotted.}
}
\description{
Plot BayesDCAList
}
\examples{
fit <- dca_list(PredModelData, cores = 4)
my_colors <- c("predictions" = "red",
               "binary_test" = "#4DAF4AFF")
plot(fit, colors = my_colors)
}
