% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dca.R
\name{plot.DiagTestDCA}
\alias{plot.DiagTestDCA}
\title{Plot DiagTestDCA}
\usage{
\method{plot}{DiagTestDCA}(obj, data_only = FALSE, .color = "#4DAF4AFF", ...)
}
\arguments{
\item{obj}{DiagTestDCA object}

\item{data_only}{If TRUE, returns data for ggplot objects (default is FALSE).}

\item{.color}{Test curve color. Default is "#4DAF4AFF" (green).}
}
\description{
Plot DiagTestDCA
}
