% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dca.R
\name{plot.PredModelDCA}
\alias{plot.PredModelDCA}
\title{Plot PredModelDCA}
\usage{
\method{plot}{PredModelDCA}(obj, type = "nb", just_df = FALSE, ...)
}
\arguments{
\item{obj}{PredModelDCA object}

\item{type}{Either "net benefit" (alias "nb" or "dca") or "parameters"
(alias "p"). If net benefit, plots estimated net benefit against
thresholds (DCA); if parameters, plots estimated prevalence,
sensitivity, and specificity.}
}
\description{
Plot PredModelDCA
}
