% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_classification}
\alias{plot_classification}
\title{Plot classification performance from Bayesian DCA of binary outcome}
\usage{
plot_classification(
  obj,
  type = c("sensitivity", "specificity"),
  strategies = NULL,
  colors = NULL,
  labels = NULL
)
}
\arguments{
\item{obj}{BayesDCA object}

\item{strategies}{Character vector with models or tests
to compare. If null, compares either first two in
\code{obj$strategies} or the first one against
Treat all/none (if only one available).}

\item{colors}{Named vector with color for each model or test. If provided
for a subset of models or tests, only that subset will be plotted.}

\item{labels}{Named vector with label for each model or test.}
}
\description{
May plot either sensitivity or specificity.
}
\keyword{internal}
