% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_delta}
\alias{plot_delta}
\title{Plot BayesDCAList delta}
\usage{
plot_delta(
  obj,
  models_or_tests = NULL,
  type = c("best", "useful", "pairwise"),
  colors = NULL,
  labels = NULL
)
}
\arguments{
\item{obj}{BayesDCAList object}

\item{colors}{Named vector with color for each model or test. If provided
for a subset of models or tests, only that subset will be plotted.}

\item{labels}{Named vector with label for each model or test.}
}
\value{
A ggplot object.
}
\description{
Plot BayesDCAList delta
}
\examples{
data(PredModelData)
fit <- dca(PredModelData, cores = 4)
plot_delta(fit)
}
