% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dca.R
\name{plot_delta_nb}
\alias{plot_delta_nb}
\title{Plot delta net benefit}
\usage{
plot_delta_nb(..., .delta = NULL, .names = NULL)
}
\arguments{
\item{...}{Pass two bayesDCA fit objects with names as they should appear in plots'
legends (e.g. "Test one" = fit1, test2 = fit2).}

\item{.delta}{Optional pre-computed delta net benefit (mostly for internal use).}

\item{.names}{Optional character vector with names for objects (mostly for internal use).}
}
\value{
A ggplot object.
}
\description{
Plot delta net benefit
}
