% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dca_list.R
\name{plot_delta_nb2}
\alias{plot_delta_nb2}
\title{Plot BayesDCAList delta}
\usage{
plot_delta_nb2(obj, models_or_tests, labels = NULL)
}
\arguments{
\item{obj}{BayesDCAList object}

\item{labels}{Named vector with label for each model or test.}

\item{colors}{Named vector with color for each model or test. If provided
for a subset of models or tests, only that subset will be plotted.}
}
\description{
Plot BayesDCAList delta
}
\examples{
fit <- dca_list(PredModelData, cores = 4)
plot_delta_nb2(fit)
plot_delta_nb2(fit, models_or_tests = "binary_test")
}
