% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_evpi}
\alias{plot_evpi}
\title{Plot Expected Value of Perfect Information (EVPI)}
\usage{
plot_evpi(
  obj,
  strategies = NULL,
  type = c("best", "useful", "pairwise"),
  colors = NULL,
  labels = NULL,
  linewidth = 1.5,
  data_only = FALSE
)
}
\arguments{
\item{obj}{BayesDCA or BayesDCASurv object}

\item{strategies}{Character vector with models or tests to plot. If null, compares either first two in \code{obj$strategies} or the first one against Treat all/none (if only one available).}

\item{type}{One of "best", "useful", or "pairwise".}

\item{colors}{Named vector with color for each model or test. If provided
for a subset of models or tests, only that subset will be plotted.}

\item{labels}{Named vector with label for each model or test.}

\item{linewidth}{Width of plotted lines.}

\item{data_only}{If TRUE, returns data.frame used for \code{ggplot2} plot with EVPI data}
}
\description{
Plot Expected Value of Perfect Information (EVPI)
}
