% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dca_list.R
\name{plot_prob_better2}
\alias{plot_prob_better2}
\title{Plot probability of better net benefit between two models or tests}
\usage{
plot_prob_better2(obj, models_or_tests, labels = NULL)
}
\arguments{
\item{obj}{BayesDCAList object}

\item{labels}{Named vector with label for each model or test.}

\item{colors}{Named vector with color for each model or test. If provided
for a subset of models or tests, only that subset will be plotted.}
}
\value{
A ggplot object.
}
\description{
Plot probability of better net benefit between two models or tests
}
\examples{
fit <- dca_list(PredModelData, cores = 4)
plot_prob_better2(fit)
plot_prob_better2(fit, models_or_tests = "binary_test")
}
