% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_superiority_prob}
\alias{plot_superiority_prob}
\title{Plot P(useful) from Wynants 2018 (doi: 10.1002/sim.7653)}
\usage{
plot_superiority_prob(
  obj,
  models_or_tests = NULL,
  type = c("best", "useful", "pairwise"),
  mcip = 0,
  colors = NULL,
  labels = NULL
)
}
\arguments{
\item{obj}{BayesDCA or BayesDCASurv object}

\item{mcip}{Minimal clinically important difference. Defaults to zero. Used only for \code{type = "pairwise"}}

\item{colors}{Named vector with color for each model or test. If provided
for a subset of models or tests, only that subset will be plotted.}

\item{labels}{Named vector with label for each model or test.}
}
\value{
A ggplot object.
}
\description{
Plot P(useful) from Wynants 2018 (doi: 10.1002/sim.7653)
}
\examples{
data(PredModelData)
fit <- dca(PredModelData, cores = 4)
plot_superiority_prob(fit)
}
