% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PredModelData}
\alias{PredModelData}
\title{Predictive Model Evaluation Data.}
\format{
A data frame with 500 rows and 2 variables:
\describe{
\item{outcomes}{binary outcomes (0 or 1, int.)}
\item{predictions}{predicted probabilities (num.)}
...
}
}
\source{
\url{https://giulianonetto.github.io/bayesdca/}
}
\usage{
PredModelData
}
\description{
A dataset containing the outcomes (0 or 1) and predictions (probabilities)
for an example model. Based on \code{rmda::dcaData}.
}
\keyword{datasets}
