% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fake_data_simulation.R
\name{simulate_prognostic_model_data}
\alias{simulate_prognostic_model_data}
\title{Simulate Prognostic Test Data}
\usage{
simulate_prognostic_model_data(N = 200, .seed = 123)
}
\description{
Simulate Prognostic Test Data
}
