% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{compare_dca}
\alias{compare_dca}
\title{Plot BayesDCA comparison}
\usage{
compare_dca(
  obj,
  strategies = NULL,
  colors = NULL,
  labels = NULL,
  plot_list = FALSE,
  .evpi = FALSE,
  type = c("best", "useful", "pairwise"),
  linewidth = 1.5,
  ...
)
}
\arguments{
\item{obj}{BayesDCA object}

\item{strategies}{Character vector with models or tests to compare. If null, compares either first two in \code{obj$strategies} or the first one against Treat all/none (if only one available).}

\item{colors}{Named vector with color for each model or test. If provided
for a subset of models or tests, only that subset will be plotted.}

\item{labels}{Named vector with label for each model or test.}

\item{plot_list}{If TRUE, returns a list of separate ggplot objects.}
}
\value{
A patchwork/ggplot object or a list of ggplot objects.
}
\description{
Plot BayesDCA comparison
}
\examples{
data(PredModelData)
fit <- dca(PredModelData, cores = 4)
compare_dca(fit)
}
