% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dca.R
\name{dca_predictive_model}
\alias{dca_predictive_model}
\title{Bayesian Decision Curve Analysis for Predictive Model}
\usage{
dca_predictive_model(
  outcomes,
  predictions,
  keep_fit = FALSE,
  prior_p = c(1, 1),
  prior_se = c(1, 1),
  prior_sp = c(1, 1),
  summary_probs = c(0.025, 0.975),
  refresh = 0,
  ...
)
}
\arguments{
\item{outcomes}{Integer vector with outcome events for each individual (0 or 1).}

\item{predictions}{Numeric vector with predicted probabilities.}

\item{keep_fit}{Logical indicating whether to keep \code{stanfit} in
the output (default FALSE).}

\item{prior_p, prior_se, prior_sp}{Numeric vector with shapes for
Beta(alpha, beta) priors used for p, Se, and Sp, respectively. Default is c(1, 1) for all (uniform prior).}

\item{refresh}{Control verbosity of \code{rstan::sampling} (check its help
page for details).}

\item{...}{Arguments passed to \code{rstan::sampling} (e.g. iter, chains).}

\item{thresholds}{Numeric vector with probability thresholds with which
the net benefit should be computed (default is \code{seq(0, 0.5, 0.01)}).}
}
\value{
An object of class \code{DiagTestDCA}
}
\description{
Bayesian Decision Curve Analysis for Predictive Model
}
\examples{
data(PredModelData)
head(PredModelData)
fit <- dca_predictive_model(outcomes = PredModelData$outcomes,
                            predictions = PredModelData$predictions)
plot(fit)
}
