% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dca_surv.R
\name{dca_surv}
\alias{dca_surv}
\title{Bayesian Decision Curve Analysis for Predictive Models and Binary tests}
\usage{
dca_surv(
  .data,
  prediction_time,
  thresholds = seq(0.01, 0.5, 0.01),
  keep_draws = TRUE,
  keep_fit = FALSE,
  summary_probs = c(0.025, 0.975),
  cutpoints = NULL,
  prior_scaling_factor = 1/3,
  use_median_surv = TRUE,
  refresh = 0,
  ...
)
}
\value{
An object of class \code{BayesDCASurv}
}
\description{
Bayesian Decision Curve Analysis for Predictive Models and Binary tests
}
\examples{
data(dca_survival_data)
fit <- dca_surv(dca_survival_data, prediction_time = 1, cores = 4)
plot(fit)
}
