% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dca_surv.R
\name{.dca_stan_surv_weibull2}
\alias{.dca_stan_surv_weibull2}
\title{Fit Bayesian Decision Curve Analysis
using Stan for survival outcomes (Weibull model)}
\usage{
.dca_stan_surv_weibull2(
  sample_size,
  n_thr,
  n_models,
  event_times_stacked,
  censored_times_stacked,
  n_event_times_stacked,
  n_censored_times_stacked,
  event_times_start_positions,
  censored_times_start_positions,
  event_times_sizes,
  censored_times_sizes,
  prediction_time,
  thresholds,
  pos_post1,
  pos_post2,
  total_event_times,
  total_censored_times,
  event_times_marginal,
  censored_times_marginal,
  other_models_indices,
  prior_only,
  prior_scale_alpha,
  prior_scale_sigma,
  iter = 2000,
  refresh = 0,
  ...
)
}
\arguments{
\item{refresh}{Control verbosity of
\href{https://mc-stan.org/rstan/reference/stanmodel-method-sampling.html}{\code{rstan::sampling}}.}

\item{...}{Arguments passed to
\href{https://mc-stan.org/rstan/reference/stanmodel-method-sampling.html}{\code{rstan::sampling}} (e.g. iter, chains).}
}
\value{
An object of class
\href{https://mc-stan.org/rstan/reference/stanfit-class.html}{\code{stanfit}} returned by \href{https://mc-stan.org/rstan/reference/stanmodel-method-sampling.html}{\code{rstan::sampling}} (e.g. iter, chains)
}
\description{
Fit Bayesian Decision Curve Analysis
using Stan for survival outcomes (Weibull model)
}
\keyword{internal}
