% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_nonconstant_prior_parameters}
\alias{.get_nonconstant_prior_parameters}
\title{Get threshold- and model-specific priors for Bayesian DCA}
\usage{
.get_nonconstant_prior_parameters(
  thresholds,
  n_models_or_tests,
  shift = 0.45,
  slope = 0.025,
  prior_sample_size = 5,
  prior_p = NULL
)
}
\arguments{
\item{n_models_or_tests}{Number of models or tests (int.).}

\item{prior_p, prior_se, prior_sp}{Non-negative shape values for
Beta(alpha, beta) priors used for p, Se, and Sp, respectively.
Default is uniform prior for all parameters - Beta(1, 1).
A single vector of the form \code{c(a, b)} can be provided for each.}

\item{n_thresholds}{Number of thresholds (int.).}
}
\description{
Get threshold- and model-specific priors for Bayesian DCA
}
\keyword{internal}
