% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dca.R
\name{get_thr_data}
\alias{get_thr_data}
\title{Get Threshold Performance Data}
\usage{
get_thr_data(outcomes, predictions, thresholds = seq(0, 0.5, 0.02))
}
\arguments{
\item{outcomes}{Integer vector (0 or 1) with binary outcomes.}

\item{predictions}{Numeric vector with predicted probabilities.}
}
\description{
Get Threshold Performance Data
}
