% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dca_list.R
\name{plot.BayesDCAList}
\alias{plot.BayesDCAList}
\title{Plot BayesDCAList}
\usage{
\method{plot}{BayesDCAList}(obj, data_only = FALSE, colors = NULL, labels = NULL, ...)
}
\arguments{
\item{obj}{BayesDCAList object}

\item{data_only}{If TRUE, returns data for ggplot objects (default is FALSE).}

\item{colors}{Named list with color for each model or test. If provided
for a subset of models or tests, only that subset will be plotted.}

\item{labels}{Named list with label for each model or test.}
}
\description{
Plot BayesDCAList
}
\examples{
fit <- dca_list(PredModelData, cores = 4)
my_colors <- list("predictions" = "red", 
                  "binary_test" = "#4DAF4AFF")
plot(fit, color = my_colors)
}
