% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.BayesDCASurv}
\alias{plot.BayesDCASurv}
\title{Plot BayesDCASurv}
\usage{
\method{plot}{BayesDCASurv}(
  obj,
  strategies = NULL,
  colors = NULL,
  labels = NULL,
  raw_values = NULL,
  raw_values_label = "Biomarker threshold",
  linewidth = 1.5
)
}
\arguments{
\item{obj}{BayesDCASurv object}
}
\description{
Plot BayesDCASurv
}
