% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dca.R
\name{plot.DiagTestDCA}
\alias{plot.DiagTestDCA}
\title{Plot DiagTestDCA}
\usage{
\method{plot}{DiagTestDCA}(obj, type = "nb", data_only = FALSE, ...)
}
\arguments{
\item{obj}{DiagTestDCA object}

\item{type}{Either "net benefit" (alias "nb" or "dca") or "parameters"
(alias "p"). If net benefit, plots estimated net benefit against
thresholds (DCA); if parameters, plots estimated prevalence,
sensitivity, and specificity.}
}
\description{
Plot DiagTestDCA
}
