% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dca.R
\name{plot.PredModelDCA}
\alias{plot.PredModelDCA}
\title{Plot PredModelDCA}
\usage{
\method{plot}{PredModelDCA}(obj, data_only = FALSE, .color = "#E41A1CFF", ...)
}
\arguments{
\item{obj}{PredModelDCA object}

\item{data_only}{If TRUE, returns data for ggplot objects (default is FALSE).}

\item{.color}{Model curve color. Default is "#E41A1CFF" (red).}
}
\description{
Plot PredModelDCA
}
\examples{
data(PredModelData)
head(PredModelData)
fit <- dca_predictive_model(outcomes = PredModelData$outcomes,
                            predictions = PredModelData$predictions)
plot(fit)
}
