% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dca.R
\name{plot_dca_list}
\alias{plot_dca_list}
\title{Merge DCA Plots}
\usage{
plot_dca_list(..., type = "nb", just_df = FALSE)
}
\arguments{
\item{...}{DiagTestDCA object}

\item{type}{Either "net benefit" (alias "nb" or "dca") or "parameters"
(alias "p"). If net benefit, plots estimated net benefit against
thresholds (DCA); if parameters, plots estimated prevalence,
sensitivity, and specificity.}
}
\description{
Merge DCA Plots
}
