% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_delta}
\alias{plot_delta}
\title{Plot BayesDCA delta}
\usage{
plot_delta(
  obj,
  strategies = NULL,
  type = c("best", "useful", "pairwise"),
  colors = NULL,
  labels = NULL,
  linewidth = 1.5
)
}
\arguments{
\item{obj}{BayesDCA or BayesDCASurv object}

\item{strategies}{Character vector with models or tests to plot. If null, compares either first two in \code{obj$strategies} or the first one against Treat all/none (if only one available).}

\item{type}{One of "best", "useful", or "pairwise".}

\item{colors}{Named vector with color for each model or test. If provided
for a subset of models or tests, only that subset will be plotted.}

\item{labels}{Named vector with label for each model or test.}

\item{linewidth}{Width of plotted lines.}

\item{plot_list}{If TRUE, returns a list of separate ggplot objects.}
}
\value{
A ggplot object.
}
\description{
Plot BayesDCA delta
}
\examples{
data(PredModelData)
fit <- dca(PredModelData)
plot_delta(fit)
}
