% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dca.R
\name{plot_delta_nb}
\alias{plot_delta_nb}
\title{Plot BayesDCAList delta}
\usage{
plot_delta_nb(obj, models_or_tests, labels = NULL)
}
\arguments{
\item{obj}{BayesDCAList object}

\item{labels}{Named vector with label for each model or test.}

\item{colors}{Named vector with color for each model or test. If provided
for a subset of models or tests, only that subset will be plotted.}
}
\value{
A ggplot object.
}
\description{
Plot BayesDCAList delta
}
\examples{
data(PredModelData)
fit <- dca(PredModelData, cores = 4)
plot_delta_nb(fit)
}
