% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_evpi}
\alias{plot_evpi}
\title{Plot Expected Value of Perfect Information (EVPI)}
\usage{
plot_evpi(
  obj,
  models_or_tests = NULL,
  type = c("best", "useful", "pairwise"),
  colors = NULL,
  labels = NULL
)
}
\arguments{
\item{obj}{BayesDCAList object}

\item{models_or_tests}{Character vector with models or tests
to compare. If null, compares either first two in
\code{obj$model_or_test_names} or the first one against
Treat all/none (if only one available).}

\item{labels}{Named vector with label for each model or test.}
}
\description{
Plot Expected Value of Perfect Information (EVPI)
}
