% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dca.R
\name{plot_prob_better}
\alias{plot_prob_better}
\title{Plot probability of better net benefit between two models or tests}
\usage{
plot_prob_better(obj, models_or_tests = NULL, labels = NULL)
}
\arguments{
\item{obj}{BayesDCAList object}

\item{models_or_tests}{Character vector with models or tests to compare. If null, compares either first two in \code{obj$model_or_test_names} or the first one against Treat all/none (if only one available).}

\item{labels}{Named vector with label for each model or test.}

\item{colors}{Named vector with color for each model or test. If provided
for a subset of models or tests, only that subset will be plotted.}
}
\value{
A ggplot object.
}
\description{
Plot probability of better net benefit between two models or tests
}
\examples{
data(PredModelData)
fit <- dca(PredModelData, cores = 4)
plot_prob_better(fit)
}
